Trying env var

+++ ddc_install.sh
@@ -1,9 +1,12 @@
 #!/bin/sh
-nodename=$1
-ucprole=$2
-ucpcontrollerip=$3
-ucpip=$4
-ucpsan=$5
+action=$1
+nodename=$2
+ucprole=$3
+ucpcontrollerip=$4
+ucpip=$5
+ucpsan=$6
+
+UCP_CONTROLLER_PROVISIONED=/tmp/ucp_controller_provisioned
 
 echo $*
 echo "-------"
@@ -11,18 +14,24 @@ echo "HOSTNAME: $(hostname)"
 echo "UCP ROLE: $ucprole"
 echo "UCP CONTROLLER IP: $ucpcontrollerip"
 echo "UCP MYIP: $ucpip"
+echo "-----------------"
+echo "ACTION: $action"
+
 
 case $ucprole in
 	controller)
+	if [ ! -f $UCP_CONTROLLER_PROVISIONED ]
+	then
 	echo docker run --rm -it --name ucp -v /var/run/docker.sock:/var/run/docker.sock \
 		docker/ucp install --host-addressdocker run --rm -it \
 		--name ucp -v /var/run/docker.sock:/var/run/docker.sock \
 		docker/ucp install --host-address $ucpip --san $ucpsan
-
+	fi
+	touch $UCP_CONTROLLER_PROVISIONED
 	;;
 
 	*)
-		echo "Undefined DDC UCP role"	
+		echo "Undefined DDC UCP role"
 	;;
 
 esac
